Extract external navigation policy handler to a new file.
